backend-service/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv # For loading .env files
firecrawl-py
openai